Move to a single Python script that implements dd
This will pro9bably be the schema for any future commands added to
gdb and lldb.
